+++ src/layout/Header/HeaderStyle.js
@@ -1,0 +1,56 @@
+import styled from 'styled-components';
+import { COLOR } from '../../assets/colors';
+export const HeaderStyle = styled.header`
+    width: 100%;
+    padding: 40px 60px 36px;
+    display: flex;
+    justify-content: end;
+    align-items: center;
+    position: relative;
+    .logo-img-wrap {
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        width: 226px;
+        img {
+            width: 100%;
+            height: 100%;
+        }
+        @media screen and (max-width: 900px) {
+            width: 184px;
+        }
+    }
+`;
+
+export const HeaderIconBox = styled.div`
+    display: flex;
+`;
+
+export const HaederButton = styled.button`
+    width: 48px;
+    height: 48px;
+    border: 1px solid ${COLOR.black};
+    background-color: ${COLOR.bg};
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+    &:nth-child(1) {
+        border-right: none;
+    }
+    svg {
+        width: 24px;
+        height: 24px;
+    }
+    &:nth-child(1) svg {
+        width: 22px;
+        height: 22px;
+    }
+    &:hover {
+        filter: brightness(0.8);
+    }
+    @media screen and (max-width: 900px) {
+        display: none;
+    }
+`;
